refactor: remove "AvailableScenarioStates" and "TestInfo" from shared types
