+++ src/compose/index.ts
@@ -1,2 +1,2 @@
-export {mapScenarios} from './map-scenarios'
-export {type TestResult} from './reporting/types'
+export { mapScenarios } from "./map-scenarios";
+export { type TestResult } from "./reporting/types";
