+++ src/modules/reusable-tests/create-reusable-test.ts
@@ -1,6 +1,6 @@
 import { ReusableTestCallback } from "./types";
 import { ReusableTest } from "./reusable-test";
-import { AvailableTestStates } from "@shared/types";
+import { AvailableScenraioStates } from "@modules/scenarios";
 
 /**
  * Creates a new test instance with the provided test logic callback.
@@ -8,7 +8,7 @@ import { AvailableTestStates } from "@shared/types";
  * @param callback The callback function containing the reusable test logic.
  * @returns A new reusable test callback.
  */
-export function createReusableTest<State extends AvailableTestStates, Params = undefined>(
+export function createReusableTest<State extends AvailableScenraioStates, Params = undefined>(
     callback: ReusableTestCallback<State, Params>
 ) {
     return new ReusableTest(callback);
