+++ src/modules/scenarios/index.ts
@@ -1,2 +1,3 @@
 export { createScenario } from "./create-scenario";
 export { Scenario } from "./scenario";
+export { type AvailableScenraioStates } from "./types";
