+++ src/modules/scenarios/scenario.ts
@@ -1,4 +1,5 @@
-import { TestInfo, TestStatus } from "@shared/types";
+import { TestInfo } from "@modules/tests";
+import { TestStatus } from "@shared/types";
 
 export class Scenario implements TestInfo {
     public status = TestStatus.NOT_RUNED;
