+++ src/modules/scenarios/types.ts
@@ -1,5 +1,7 @@
 import { TestFactory } from "@modules/tests";
 
+export type AvailableScenraioStates = Record<string, unknown> | null;
+
 type ScenarioCallbackPayload<State> = {
     test: TestFactory<State>;
 };
