+++ src/modules/tests/classes/case-test.ts
@@ -1,5 +1,6 @@
 import { TestAborter } from "@core/aborting";
-import { TestInfo, TestStatus } from "@shared/types";
+import { TestStatus } from "@shared/types";
+import { TestInfo } from "../types";
 
 type Callback = () => void | Promise<void>;
