+++ src/modules/tests/factories/test-factory.ts
@@ -1,7 +1,7 @@
 import { LifecycleHookCallback } from "@modules/lifecycle-hooks";
 import { GroupNode } from "@modules/testing-tree";
 import { abort } from "@core/aborting";
-import { SampleTestCallback, TestsGroupCallback } from "../types/callbacks";
+import { SampleTestCallback, TestsGroupCallback } from "../types";
 import { TestCase } from "../classes/case-test";
 import { TestsGroup } from "../classes/test-group";
 import { BaseTestFactory } from "./base-test-factory";
