+++ src/modules/tests/index.ts
@@ -2,3 +2,4 @@ export { TestCase } from "./classes/case-test";
 export { TestsGroup } from "./classes/test-group";
 export { BaseTestFactory } from "./factories/base-test-factory";
 export { TestFactory } from "./factories/test-factory";
+export { type TestInfo } from "./types";
